feat(tools): expose GeoUI Protocol tools over MCP
  Move the GeoUI Protocol (core schema, Worldview adapter, URL parser,
  extensions) into akd_ext/tools/geoui and register the render-intent
  and get-state tools with @mcp_tool, since only akd_ext* is packaged.
  ieso_w_geoui keeps working via re-export shims; round-trip and
  registration tests added.
